moved BaseFeaturizer to new file
